@@ -19,22 +19,22 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering :: Astronomy",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-    "numpy>=1.18.5",
-    "astropy>=4.0,!=4.0.1,!=4.0.1.post1",
+    "numpy>=1.23.0",
+    "astropy>=5.0.5",
     "pyerfa",
-    "scipy>=1.4.1",
+    "scipy>=1.9.0",
     "jplephem>=2.6",
-    "matplotlib>=3.2.0",
+    "matplotlib>=3.4.3",
     "emcee>=3.0.1",
     "corner>=2.0.1",
     "uncertainties",
