@@ -83,14 +83,25 @@ def main(argv=None):
     parser.add_argument(
         "-q", "--quiet", default=0, action="count", help="Decrease output verbosity"
     )
+    parser.add_argument(
+        "--allow_tcb",
+        action="store_true",
+        help="Convert TCB par files to TDB automatically",
+    )
+    parser.add_argument(
+        "--allow_T2",
+        action="store_true",
+        help="Guess the underlying binary model when T2 is given",
+    )
 
     args = parser.parse_args(argv)
     pint.logging.setup(
         level=pint.logging.get_level(args.loglevel, args.verbosity, args.quiet)
     )
 
-    m1 = get_model(args.input1)
-    m2 = get_model(args.input2)
+    m1 = get_model(args.input1, allow_T2=args.allow_T2, allow_tcb=args.allow_tcb)
+    m2 = get_model(args.input2, allow_T2=args.allow_T2, allow_tcb=args.allow_tcb)
+
     print(
         m1.compare(
             m2,
