@@ -128,6 +128,17 @@ def main(argv=None):
         help="Logging level",
         dest="loglevel",
     )
+    parser.add_argument(
+        "--allow_tcb",
+        action="store_true",
+        help="Convert TCB par files to TDB automatically",
+    )
+    parser.add_argument(
+        "--allow_T2",
+        action="store_true",
+        help="Guess the underlying binary model when T2 is given",
+    )
+
     global nwalkers, nsteps, ftr
 
     args = parser.parse_args(argv)
@@ -164,7 +175,9 @@ def main(argv=None):
     wgtexp = args.wgtexp
 
     # Read in initial model
-    modelin = pint.models.get_model(parfile)
+    modelin = pint.models.get_model(
+        parfile, allow_T2=args.allow_T2, allow_tcb=args.allow_tcb
+    )
 
     # The custom_timing version below is to manually construct the TimingModel
     # class, which allows it to be pickled. This is needed for parallelizing
