@@ -74,6 +74,16 @@ def main(argv=None):
     parser.add_argument(
         "-q", "--quiet", default=0, action="count", help="Decrease output verbosity"
     )
+    parser.add_argument(
+        "--allow_tcb",
+        action="store_true",
+        help="Convert TCB par files to TDB automatically",
+    )
+    parser.add_argument(
+        "--allow_T2",
+        action="store_true",
+        help="Guess the underlying binary model when T2 is given",
+    )
 
     args = parser.parse_args(argv)
     pint.logging.setup(
@@ -105,7 +115,9 @@ def main(argv=None):
     )
 
     if args.parfile is not None:
-        m = pint.models.get_model(args.parfile)
+        m = pint.models.get_model(
+            args.parfile, allow_T2=args.allow_T2, allow_tcb=args.allow_tcb
+        )
     else:
         # Construct model by hand
         m = pint.models.StandardTimingModel
